Types mistakes

+++ circus/clustering.py
@@ -733,7 +733,7 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
                 cfile      = h5py.File(file_out_suff + '.clusters.hdf5', 'r+', libver='latest')
                 templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*n_clusters), dtype=numpy.float32, chunks=True)
                 electrodes = hfile.create_dataset('electrodes', shape=(n_clusters, ), dtype=numpy.int32, chunks=True)
-                norms      = hfile.create_dataset('norms', shape=(2*n_clusters, ), dtype=numpy.int32, chunks=True)
+                norms      = hfile.create_dataset('norms', shape=(2*n_clusters, ), dtype=numpy.float32, chunks=True)
                 amplitudes = hfile.create_dataset('limits', shape=(n_clusters, 2), dtype=numpy.float32, chunks=True)
                 count      = 0
                 for i in xrange(comm.size):
@@ -895,7 +895,7 @@ def main(filename, params, nb_cpu, nb_gpu, use_gpu):
                 cfile      = h5py.File(file_out_suff + '.clusters.hdf5', 'w', libver='latest')
                 templates  = hfile.create_dataset('templates', shape=(N_e, N_t, 2*n_clusters), dtype=numpy.float32, chunks=True)
                 electrodes = hfile.create_dataset('electrodes', shape=(n_clusters, ), dtype=numpy.int32, chunks=True)
-                norms      = hfile.create_dataset('norms', shape=(2*n_clusters, ), dtype=numpy.int32, chunks=True)
+                norms      = hfile.create_dataset('norms', shape=(2*n_clusters, ), dtype=numpy.float32, chunks=True)
                 amplitudes = hfile.create_dataset('limits', shape=(n_clusters, 2), dtype=numpy.float32, chunks=True)
                 count      = 0
                 for i in xrange(comm.size):
